requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.1.0
flask-login>=0.6.0
werkzeug>=2.3.0
pymysql>=1.1.0
sentence-transformers>=2.2.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
pdfplumber>=0.10.0
cryptography>=41.0.0